+++ ares/n64/memory/bus.hpp
@@ -18,11 +18,7 @@ inline auto Bus::read(u32 address) -> u64 {
   if(address <= 0x048f'ffff) return si.read<Size>(address);
   if(address <= 0x04ff'ffff) return unmapped;
   if(address <= 0x1fbf'ffff) return pi.read<Size>(address);
-  if(address <= 0x1fc0'07bf) {
-    if(pi.io.romLockout) return unmapped;
-    return pi.rom.read<Size>(address);
-  }
-  if(address <= 0x1fc0'07ff) return pi.ram.read<Size>(address);
+  if(address <= 0x1fcf'ffff) return pif.read<Size>(address);
   return unmapped;
 }
 
@@ -49,10 +45,6 @@ inline auto Bus::write(u32 address, u64 data) -> void {
   if(address <= 0x048f'ffff) return si.write<Size>(address, data);
   if(address <= 0x04ff'ffff) return;
   if(address <= 0x1fbf'ffff) return pi.write<Size>(address, data);
-  if(address <= 0x1fc0'07bf) {
-    if(pi.io.romLockout) return;
-    return pi.rom.write<Size>(address, data);
-  }
-  if(address <= 0x1fc0'07ff) return pi.ram.write<Size>(address, data);
+  if(address <= 0x1fcf'ffff) return pif.write<Size>(address, data);
   return;
 }
