+++ ares/n64/n64.hpp
@@ -66,6 +66,7 @@ namespace ares::Nintendo64 {
   #include <n64/vi/vi.hpp>
   #include <n64/ai/ai.hpp>
   #include <n64/pi/pi.hpp>
+  #include <n64/pif/pif.hpp>
   #include <n64/ri/ri.hpp>
   #include <n64/si/si.hpp>
   #include <n64/rdram/rdram.hpp>
