+++ ares/n64/pi/debugger.cpp
@@ -1,13 +1,4 @@
 auto PI::Debugger::load(Node::Object parent) -> void {
-  memory.ram = parent->append<Node::Debugger::Memory>("PI RAM");
-  memory.ram->setSize(64);
-  memory.ram->setRead([&](u32 address) -> u8 {
-    return pi.ram.read<Byte>(address);
-  });
-  memory.ram->setWrite([&](u32 address, u8 data) -> void {
-    return pi.ram.write<Byte>(address, data);
-  });
-
   tracer.io = parent->append<Node::Debugger::Tracer::Notification>("I/O", "PI");
 }
