+++ ares/n64/pi/serialization.cpp
@@ -1,6 +1,4 @@
 auto PI::serialize(serializer& s) -> void {
-  s(ram);
-
   s(io.dmaBusy);
   s(io.ioBusy);
   s(io.error);
@@ -9,7 +7,6 @@ auto PI::serialize(serializer& s) -> void {
   s(io.pbusAddress);
   s(io.readLength);
   s(io.writeLength);
-  s(io.romLockout);
   s(io.busLatch);
 
   s(bsd1.latency);
