+++ ares/n64/pif/serialization.cpp
@@ -1,0 +1,5 @@
+auto PIF::serialize(serializer& s) -> void {
+  s(ram);
+
+  s(io.romLockout);
+}
