+++ ares/n64/si/dma.cpp
@@ -1,5 +1,5 @@
 auto SI::dmaRead() -> void {
-  run();
+  pif.run();
   for(u32 offset = 0; offset < 64; offset += 2) {
     u16 data = bus.read<Half>(io.readAddress + offset);
     bus.write<Half>(io.dramAddress + offset, data);
@@ -17,5 +17,5 @@ auto SI::dmaWrite() -> void {
   io.dmaBusy = 0;
   io.interrupt = 1;
   mi.raise(MI::IRQ::SI);
-  run();
+  pif.run();
 }
